Language class og små rettelser

+++ 24_CDIO_Final/src/fields/Fleet.java
@@ -88,8 +88,8 @@ public class Fleet extends Ownable {
 		return pawned;
 	}
 
-	public double getPawnPrice() {
-		double pawnprice = ((int) fieldprice * 0.5);
+	public int getPawnPrice() {
+		int pawnprice = (int) (fieldprice * 0.5);
 		return pawnprice;
 	}
