+++ 24_CDIO_Final/src/fields/Labor.java
@@ -110,8 +110,8 @@ public class Labor extends Ownable {
 		return pawned;
 	}
 
-	public double getPawnPrice() {
-		double pawnprice = ((int) fieldprice * 0.5);
+	public int getPawnPrice() {
+		int pawnprice = (int) (fieldprice * 0.5);
 		return pawnprice;
 	}
